+++ TelCo.ColorCoder/GetColorFromPairNumber.cs
@@ -1,0 +1,41 @@
+using System;
+using System.Diagnostics;
+using System.Drawing;
+
+namespace TelCo.ColorCoder
+{
+    /// <summary>
+    /// The 25-pair color code, originally known as even-count color code, 
+    /// is a color code used to identify individual conductors in twisted-pair 
+    /// wiring for telecommunications.
+    /// This class provides the color coding and 
+    /// mapping of pair number to color and color to pair number.
+    /// </summary>
+    partial class Program
+    {
+        private static ColorPair GetColorFromPairNumber(int pairNumber)
+        {
+            // The function supports only 1 based index. Pair numbers valid are from 1 to 25
+            int minorSize = colorMapMinor.Length;
+            int majorSize = colorMapMajor.Length;
+            if (pairNumber < 1 || pairNumber > minorSize * majorSize)
+            {
+                throw new ArgumentOutOfRangeException(
+                    string.Format("Argument PairNumber:{0} is outside the allowed range", pairNumber));
+            }
+            
+            // Find index of major and minor color from pair number
+            int zeroBasedPairNumber = pairNumber - 1;
+            int majorIndex = zeroBasedPairNumber / minorSize;
+            int minorIndex = zeroBasedPairNumber % minorSize;
+
+            // Construct the return val from the arrays
+            ColorPair pair = new ColorPair() { majorColor = colorMapMajor[majorIndex],
+                minorColor = colorMapMinor[minorIndex] };
+            
+            // return the value
+            return pair;
+        }
+        
+    }
+}
